i#5843 scheduler: Implement speculation with nops (#6016)
Implements initial speculation support, supplying nops. Speculation is
separated into its own class where we can fill in different strategies
in the future.

The start_speculation() function takes a flag controlling whether the
scheduler queues up the current record and re-returns it as the first
record after speculation stops. This is often what a simulator wants as
it has to read the instruction record following a branch to determine
whether it is on the wrong path, and it would like to resume with that
already-read instruction after speculation.

Adds a unit test.

Issue: #5843