i#5858: Fix drmemtrace docs location in menu
The online html docs ended up with the Tracing and Analysis Framework
at the top level listed under License, rather than a sub-menu entry of
Available Tools, due to PR #5685 which moved drcachesim.dox.in into a
docs/ subdir and now the cmake glob doesn't find it.

We fix that by extending the glob.

Tested with a local docs build.

Fixes #5858