i#5843 scheduler: Update checked-in trace (#6038)
Updates the checked-in thread sig trace, which is old and small, to a
new trace in the zip format and with more threads and more code. This is
a better stress test of the scheduler. This trace is already used by the
scheduler tests.

Augments the scheduler_launcher to add a sanity test running it on the
new trace added here, and improves a few other things identified in PR #6039.

Issue: #5843