iX: Expand sizes of non-contiguous or offset half-width operands
Encoding the unpack-and-interleave instructions and vmovddup with the
half-width operands is not in general correct. While the *number*
of bytes taken from the operand is correct, those bytes are not in
general taken in a contiguous sequence starting from offset zero
within the operand. A tool using this operand information will
capture the wrong instruction inputs.

movddup, unpcklps, and unpcklpd actually do degenerate to true
half-width operands, as they start at offset zero and only use the
first eight bytes. The other opcodes, including the AVX versions
of however, use bytes that are offset from zero, that are non-contiguous,
or both. Their representation is changed to be full-width.

For prior art see e.g. movsldup/movshdup which also take a
non-contiguous set of bytes and are encoded with full-width operands.