drcachesim: added use_physical option to config file reader (#6051)
As the title says, this makes an existing option (use_physical)
available in the cache simulator config file reader, so that it can be
used when a config file is used to initialise the cache simulator
instead of command line options.

Adds a test to tests/config_reader_test.cpp and links that test file
up to drcachesim_unit_tests: it was added in commit fa4cc4a but
was never actually built or run!  Now it is built and tested.