i#6020 interval analysis: Add active shard count
Adds count of shards active in the interval to analysis_tool_t::interval_state_snapshot_t.
This is another field that may be convenient for tools to have.

Adds to the print_interval_results() doc that tools should avoid free-ing memory or
modifying global state, so that the analyzer framework can invoke it multiple times. This
is to keep open the possibility of the framework printing interval results for each
shard separately.

Adds missing comparison for instr_count_delta and instr_count_cumulative to
recorded_snapshot_t::operator==();

Issue: #6020