i#6020 interval analysis: Allow multiple print_interval_results calls
Adds to the print_interval_results() doc that tools should avoid free-ing memory
or modifying global state, so that the analyzer framework can invoke it multiple
times. This is to keep open the possibility of the framework printing interval
results for each shard separately in the future by invoking this API multiple
times.

Adds missing comparison for instr_count_delta and instr_count_cumulative to
recorded_snapshot_t::operator==().

Issue: #6020