i#2204: Retry flaky tests in suite (#6075)
If cmake 3.17+ is in use, enables retrying of failed tests up to 3x with
any one of them passing resulting in an overall pass. This avoids flaky
tests marking the whole suite red, which is even more problematic with
merge queues.

All of our Github Actions platforms (macos-11, windows-2019,
ubuntu-20.04, ubuntu-22.04) have cmake 3.26+ so this is enabled for all
of our GA CI tests.

Tested:
I made a test which fails 3/4 of the time:
```
  add_test(bogus bash -c "exit \$((RANDOM % 4))")
```
I made it easy to run just this test (gave it a label; disabled other
builds, etc.) for convenience and then ran:
```
  $ ctest -VV -S ../src/suite/runsuite.cmake,64_only
```
Which resulted in:
```
  test 1
      Start 1: bogus

  1: Test command: /usr/bin/bash "-c" "exit $((RANDOM % 4))"
  1: Working Directory: /usr/local/google/home/bruening/dr/git/build_suite/build_debug-internal-64
  1: Test timeout computed to be: 600
  1/1 Test #1: bogus ............................***Failed    0.00 sec
      Start 1: bogus

  1: Test command: /usr/bin/bash "-c" "exit $((RANDOM % 4))"
  1: Working Directory: /usr/local/google/home/bruening/dr/git/build_suite/build_debug-internal-64
  1: Test timeout computed to be: 600
      Test #1: bogus ............................   Passed    0.00 sec

  100% tests passed, 0 tests failed out of 1
```

Issue: #2204, #5873
Fixes #2204