i#6020 interval analysis: add TID and positive interval_id (#6082)
Adds TID to the interval_state_snapshot_t base struct, so that tools do
not need to track it themselves if they need it. The TID is stored as
"shard_id" which in future may contain a different id if we define
shards differently.

Also changes interval_id to always be a positive integer, which makes
some things simpler downstream.

Issue: #6020