i#5843 scheduler: Add syscall number markers
Adds a new marker TRACE_MARKER_TYPE_SYSCALL whose value holds the
system call number.  This will be used by the scheduler for
maybe-blocking syscalls (the plan is to have raw2trace add this, in a
separate PR) as well as for identifying where to insert canned kernel
syscall sequences.

Adds a new filetype bit OFFLINE_FILE_TYPE_SYSCALL_NUMBERS which
indicates whether these markers are present.

Adds 3 new checks to the invariant_checker verifying that these
markers are in the right places, which automatically adds tests of
this to our various existing regression tests.

Issue: #5843