Adds removal of syscall instrs without number markers in raw2trace.
We have 2 scenarios where we record a syscall instr yet it doesn't
execute right away and so there's no syscall number marker:
1) An asynchronous signal arrives during the block and is delivered
   from dispatch before handling the syscall.
2) A syscall ends a tracing window and we disable tracing before the
   syscall is handled.
In both cases, we want to just remove the syscall instr: so we remove
if we find no subsequent marker either immediately following or after
a buffer header of timestamp+cpuid.
(For the window case there are alternatives where we try to emit
the marker by passing info to the pre-syscall event handler or by moving
the marker to the block instrumentation but these all incur more complexity
than this relatively simple solution.)

Adds a new test linux.signal_pre_syscall which manages to get a signal
to arrive while inside a syscall-terminated block.  Adds a drcacheoff
test using this app: it is a little slow but if we shrink the app the
scenario is less likely to reproduce.