initialize kernel_sigaction_t variables in signal.c
When building dynamorio with -Wmaybe-uninitialized, the compiler
reported the error. At a first glance, it looks like this is a false
positive, as the variables are initialized by either of the above
branches. However, it does not hurt to initialize them to make the
compiler happy.

This is also helps when packaging for distros, as there usually this
warning is enabled.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>