i#6107: Fix schedule file count error
Fixes an error in raw2trace where the instruction count it writes in
the schedule files is modulo the chunk count.

Adds checks in the scheduler where it raises an error on a decreasing
count.

Adds a workaround where the scheduler detects a schedule file with the
bug and corrects the counts.  Adds a unit test for this.  I also
tested this on a real, large schedule file from the
linux.signal-pre-syscall test which is part of PR #6096.

Relaxes the invariant check on an empty schedule file, which happens
when zip support is not linked in.  Maybe it would be better to change
raw2trace to write out at least the serial file in this case but it
does not seem worth the effort to refactor the code for this build
setup where fast skipping would not be supported in any case.  This
relaxation is tested by the forthcoming drcacheoff.sysnums test in PR

Issue: #5843, #6107
Fixes #6107