i#6020 interval analysis: Fix interval counts regex (#6116)
Ensures that we do not track unique PCs in the combined interval
snapshots. Since the shard-local snapshots themselves do not track
unique pcs, the combined snapshot also wouldn't have anything in
the unique pc set. But we need to explicitly disable unique pcs
tracking so that we skip printing that field. This was missing in the
original PR #6097.

Also fixes expected regex of basic counts interval output so that only
digits and leading white-spaces are accepted at the beginning of each
line. Otherwise we aren't able to detect that the unique pc addrs line
in the counters output is still present.

Also fixes a regression due to #6113 which causes a regex match
failure in case only a single interval is output for the whole trace.

Issue: #6020, #6112