i#6107 sched count bug: Fix schedule file count error (#6114)
Fixes an error in raw2trace where the instruction count it writes in the
schedule files was incorrectly modulo the chunk count.

Adds checks in the scheduler to raise an error on a decreasing count.

Adds a workaround where the scheduler detects a schedule file with the
bug and corrects the counts. This handles multiple inputs by building
and sorting per-input sequences and then updating the schedule
structures in a separate
step using a timestamp-keyed hashtable.

Adds a unit test for the workaround. I also tested this on a real, large
schedule file from the linux.signal-pre-syscall test which is part of PR
#6096.

Fixes the scheduler to support replaying as-traced with a single output
if that is what the user requests (I needed this for the just-mentioned
manual test on the pre-syscall trace).

Disables writing any serial schedule file without zip support. The
raw2trace code was already conditional on zip support, so we were
creating an empty file. The serial file is meant only for fast skipping
support, which requires zip format.
This fixes an invariant check failure in the forthcoming
drcacheoff.sysnums test in PR #6096 from an emtpy file.

Issue: #5843, #6107