i#6020: add raw2trace stat for trace duration.
Adds raw2trace stats for earliest and latest trace timestamp. These are used
to compute the trace's wall-clock duration which is printed at the end of
raw2trace decomposition, along with other raw2trace stats.

It may be useful for the user to know the trace's wall-clock duration to
compute the value for -trace_interval (which is the length of each trace
interval for interval analysis) to get the required number of total intervals.

Issue: #6020