i#4343: Add dynamorio::drmemtrace namespace to all drcachesim/ code
Adds the dynamorio::drmemtrace namespace to all drcachesim/ code.  It
was already present on newer code.

This breaks compatibility and external code using these files will
need to be updated and recompiled.

Shifts suite/tests/rseq.c to be C++ code so it can continue to test
drmemtrace_client_main() which is now inside the namespace.

Fixes #4343