i#3044 AArch64 SVE codec: Fix LDR/STR <P> mem size (#6148)
The svemem_gpr_simm9_vl operand was setting the memory transfer size to
the current vector length, which is correct for Z register loads and
stores but not P register loads and stores.

The encoder/decoder will now detect which kind of register is being
accessed and set the correct size.

Issues: #3044