i#5843 scheduler: Switch on a blocking syscall
Adds an input switch on a blocking syscall, without considering the
current input even if it is the highest priority.

Refactors pick_next_input() to use a helper function
pick_next_input_as_previously() to make it easier to read the new wait
state code.

Adds a unit test with blocking syscalls and priorities to ensure
lower-priority inputs are scheduled when a higher-priority input goes
into a wait state.

Regenerates the checked-in threadsig.x64 trace to contain syscall
number and maybe-blocking markers and updates the associated tests'
thread ids and expected output.

Issue: #5843