i#6104 multi-window: Refresh timestamps in a new window
Updates the timestamp in the first buffer in a new window in
multi-window traces. This fixes an issue where mult-window
traces would have low activity at the beginning of a new window.

Adds a unit test that fails without the fix in this change but passes
with the change.

Fixes #6143