Fix i-filtered trace issues
The #4343 namespace change PR #6169 changed suite/tests/rseq.c to be C++ which somehow triggered two problems with i-filtered traces.  To separate the fixes for those from the large namespace change we commit them separately:

+ Do not try to roll back rseq for i-filtered traces.

+ Do not complain about signal placement vs test-app annotations for
  i-filtered traces.

We rely on the rseq-filter test in PR #6169 as a test and do not
attempt to add a different test here.

Issue: #4343