Sort also by CPU for schedule file invariant check.
In some internal tests we modify the TID of some entries to test
the schedule file invariant check. This inadvertently inverts the
order of some entries for this check if the new synthetic TID
matches with an existing real TID. By sorting also by CPU id, we
ensure consistent ordering.