i#5844 Move x86 expand_scatter_gather to its own file (#6173)
The existing drx_expand_scatter_gather() function is > 2000 lines of
code, including helper functions, and most of it is very x86 specific.

In preparation for writing the AArch64 version of this function, this
commit moves the x86 implementation of drx_expand_scatter_gather() out
of drx.c into its own architecture specific files, and creates dummy
implementations for other architectures.

The behaviour of the function should be unchanged on all platforms.

As we develop the AArch64 version of drx_expand_scatter_gather(), any
parts of the original x86 implementation which are usefull to both can
be pulled out into helper functions which are shared between both
architectures.

Issues: #5844, #5036, #3837