i#4343 namespaces: Shrink Doxygen qualifications where possible
By adding Doxygen docs to namespace declarations we can shrink many of
the awkwardly long references in Doxygen comments.  Some still produce
errors if not qualified which I do not understand so those are left.
I only updated a couple of key headers which contained the worst
instances.

This completes the namespace cleanup.

Fixes #4343