i#6196: fix UBSAN-reported issue in opnd_type_ok
di->final_pc is NULL (not initialized) when called from encoding_possible().
So the reachability check does not really work and it's not necessary
to determine possible encodings.
Adding something to a NULL pointer is also an UB as detected by UBSAN.
So remove the reachability check in opnd_type_ok() for TYPE_O.

Fixes: #6196