i#6196: Avoid adding to a null pointer
Avoids non-regular-encode code paths which do not set a final PC and
thus have a null PC value from manipulating that PC in opnd_type_ok().

Documents that get_encoding_info() may not find the same template as
when called with the real final PC.  Documents in key callers that
this is fine.

Adjusts the raw2trace unit test helper to use a non-null final PC.

Fixes #6196