drcachesim: add cache name and parameter accessors. (#6200)
Adds a name to caching_device, with accessors for the name and other
cache parameters. Added a "caching_device" pointer to
caching_device_stats, to make it easier for client code to access the
cache associated with a stats object. Updated drcachesim output reports
to include cache names and a short description of each cache's
configuration. Updated tests for the output format changes and new
methods.