RFC: convert global opcode string pointers to offsets
The goal of this change is to reduce drdecode library size.
Static library size is reduced by ~200KB (18%) and
runtime VM footprint is reduced by ~50KB (10%).

    FILE SIZE        VM SIZE
 --------------  --------------
  +139% +36.2Ki  +140% +36.2Ki    .rodata
  +0.4%    +100  [ = ]       0    .strtab
  +6.1%     +79  [ = ]       0    .comment
  +0.2%     +72  [ = ]       0    .symtab
  +5.9%     +64  [ = ]       0    [ELF Header]
  +4.7%     +60  [ = ]       0    [AR Headers]
  +0.1%     +48  [ = ]       0    .rela.text
  +7.3%      +4  [ = ]       0    .llvm_addrsig
  -0.2%    -220  -0.2%    -211    .text
  -0.7%    -576  [ = ]       0    [ELF Section Headers]
 -87.2% -28.2Ki -87.2% -28.2Ki    .rodata.str1.1
 -19.8% -62.9Ki -19.8% -62.9Ki    .data.rel.ro
 -63.5%  -151Ki  [ = ]       0    .rela.data.rel.ro
 -18.6%  -206Ki -10.5% -55.1Ki    TOTAL

Any pointers in .data/.rodata take additional 24 bytes
in the binary for a relocation (on top of the pointer and
pointee themselves).
This change replaces opcode names in decoding tables with
offsets into a specially created string that contains all opcodes.
This removes the 24-byte relocation and reduces 8-byte pointer
with a 2-byte offset (it takes the alignment padding
in instr_info_t, so effectively -8 bytes).

The idea behind the approach is to keep the source code
as readable and as maintainable as possible, while still
replacing direct string literal pointers with offsets.

The crux of the approach is as follows.
A string "foo" is replaced with STROFF(foo).
A code generator extracts all STROFF(.*) from source files
and generates a single global concatenated string that contains
all individual strings. Plus it generates STROFF_foo macro that
expands to the offset of the string "foo" in the global string.
With the addition of:
	define STROFF(foo) STROFF_ ## foo
the source code does not need any additional modifications.

The downside is that strings cannot contain any characters
that cannot appear in C identifiers. This is not a problem
for real opcodes, but affects some fake internal names
(e.g. "(bad)" is converted to "BAD").

If this change looks good in general, I will work on productionizing it:
 - convert other arches
 - integrate code generator into the build system
 - rewrite the script in Python

Also:
 - does the STROFF prefix for everything looks good?
   should I change it to something else?
 - suggestions for conversion rules for current strings
   that are not valid C identifiers are welcome
 - potentially all strings can be kept as is, if we do:
	STROFF(bad, "(bad)")
   normal opcodes can still be STROFF(mov),
   but not sure if it's worth it