i#725: Add an automated test of detach on Windows (#6221)
This patch adds an automated test of detach on Windows platform. The
test start a target program and attach to the process firstly, then use
the drconfig front-end to detach from the process.
As shown in the .template file, the detach mechanism lets the target
process be detached before stopping it.

Issue: [#725](https://github.com/DynamoRIO/dynamorio/issues/725)

---------

Co-authored-by: Derek Bruening <bruening@google.com>