i#5843 scheduler: Delay switch to instr after blocking syscall
Fixes a problem with the implementation of context switching on a
blocking syscall: it was being done immediately upon seeing the
maybe-blocking marker, which is right after the syscall instruction.
This resulted in recording the syscall instruction ordinal as the
switch point, when that ordinal is supposed to be exclusive and so
needs to be right before the subsequent instruction.

Adds a test that ensures we don't switch until after all the markers
after a blocking syscall.

Issue: #5843