i#5843 scheduler: Fix record-replay ordinal mismatches
Adds read-ahead to timestamps when replaying, to match ordinals when recording.

Adds use of the just-read timestamp when skipping in the reader, to
avoid losing the timestamp when the scheduler has read ahed.

Adds recording of the ordinals at switch points in the unit tests for
comprehensive testing of switch points instead of just testing thread
sequences.

Fixes several tests which were passing incorrect flags.

Issue: #5843