i#5490 ind br tgt: Stop skip and switch at target marker
Updates reader_t::skip_instructions() and the scheduler's context
switches and replay to stop at an indirect branch target marker
instead of at the subsequent instruction record itself.

Adds an indirect branch to allasm_x86 and uses it in a new
skip_unit_tests test.

Adds a new raw2trace test that an indirect branch marker is after a
chunk boundary.

Adds scheduler tests that a context switch during live and replay
modes goes to the marker.

Updates the invariant checker to avoid failure on the offline.skip*
tests where the filetype record is not seen: now the checker uses the
stream to get the filetype, except for the checker unit tests (added a
XXX comment on possibly giving them a valid stream).

Issue: #5490