i#5843 scheduler: Add syscall marker counts to basic_counts (#6269)
Adds the count of system call number markers and blocking system call
markers to the basic_counts tool. Updates the output in the docs and key
tests. Removes the long marker list from most tests which are not really
checking any values anyway, to simplify future additions to the marker
types counted.

Issue: #5843