i#6131: Properly disable Windows popups
Adds debug flags to the test_helpers library to actually disable
Windows popups as the prior attempt actually only succeeded on a
couple of executables.

Confirmed by running dumpbin on each binary and ensuring its
disable_popups() function calls _CrtSetReportMode.  This is rather
slow though so it's not clear we want to add this as a build check.

Removes test ignore directives listing #6131.  (The gencode one is
actually most likely properly fixed by PR #6281 as it was probably the

Issue: #6280, #6131
Fixes #6131