i#6104 multi-window: Refresh timestamps in a new window (#6165)
Updates the timestamp in the first buffer of a new window in
multi-window traces
by refreshing the timestamp when re-starting tracing. Previously, the
timestamp in
the first buffer was coming from the end of a prior window causing low
activity at the
beginning of a new window.

Adds a unit test that fails without the fix in this change but passes
with the change.

Fixes an issue where multi-window tracing fails to write the last window
raw file.

Fixes #6104 
Fixes #6245