i#6288: Handle function markers delayed by signal arriving after a branch.
When a signal arrives after a branch to a function, function markers
are inserted after the corresponding syscall xfer marker.

There are a few corner cases which need to be handled:
1. Function markers in the beginning of a trace before any instructions.
2. syscall xfer markers without the corresponding kernel xfer markers.
   This case happens when a trace is started in the middle of a signal
   handler, or nested signals.

In order to avoid false positives, the check for function markers after
a branch assumes markers are place correctly in these cases.