i#3544 RV64: Assert AUIPC instruction imm alignment (#6254)
Adds an assert to `encode_u_immpc_opnd` to ensure that the lower 12 bits
of imm are all 0s.

Also, in order to share `.expect` files between static link and dynamic
link tests, we added the ability to ignore specific lines in
runcmp.cmake and removed the dependence on the external `diff` command.
Note that we fall back to `STREQUAL` when the expected file is too
large. We can use ZIP_LISTS to provide better performance, but this
feature requires CMake 3.17, but bumping to 3.17 requires some effort on
the Windows side.

Related: #6208 
Issue: #3544