i#6294: Handle consecutive window-final branches
Adds a test case for consecutive window-final branches which
reproduces a problem where the final is removed but the penultimate is
then not marked taken/untaken.

Updates raw2trace to correctly mark the penultimate.  Also confirmed
this fix on the drcacheoff.windows-invar test on aarch64 where it was
first observed.

Fixes #6294