i#5843 scheduler: Add unread_last_record()
Adds a new scheduler output stream interface unread_last_record().
This queues the just-read record for reading later.  This is meant to
aid simulators reading one record ahead to the next instruction, but
not yet processing it.  The new interface is limited in that it can
only be called once at a time and while the output ordinals are fixed
up other metrics like the output last timestamp and the input ordinals
are not.

Adds some sanity tests.

Issue: #5843