i#5843 scheduler: Add unread_last_record() (#6302)
Adds a new scheduler output stream interface unread_last_record(). This
queues the just-read record for reading later. This is meant to aid
simulators reading one record ahead to the next instruction, but not yet
processing it. The new interface is limited in that it can only be
called once at a time and while the output ordinals are fixed up other
metrics like the output last timestamp and the input ordinals are not.

Adds a new record type TRACE_TYPE_INVALID for use as a sentinel.

Adds some sanity tests.

Issue: #5843