i#5036 Fix memval_simple_scattergather failures
While developing the the AArch64 version of drx_expand_scatter_gather
I have seen the code_api|sample.memval_simple_scattergather fail
intermittently with the not yet implemented assert from
drx_scatter_gather_restore_state() and tracked down the problem to
the trace buffer management mechanism. When the trace buffer is full
the next store will be past the end of the buffer and trigger a fault.
The fault is meant to be intercepted by the client and used to flush
the buffer.

If the buffer happens to be full when tracing an expanded scatter/gather
instruction then the scatter/gather state restore function will be
called so it would previously just fail with the assert because state
restoration is not implemented yet.

Until scatter/gather state restore on AArch64 is implemented properly we
can work around the problem by marking the expansion loads and stores we
emit with a note that we can check for in the state restore function.
If the faulting access does not originate from an expansion instruction
then we pass the fault on to the client to handle instead.

Issue: #5036