i#6303: Avoid duplicate encoding entry on chunk-split branch pair
Fixes a case where a branch followed by an indirect branch with a
chunk split in the middle results in an extra encoding entry after the
indirect branch target marker.  This extra entry is accumulated in the
reader and results in a fatal error.

Adds a unit test.  Confirmed the fix on the
drcacheoff.invariant_checker test as well where it was failing
reliably on one machine.

Fixes #6303