i#4465: Trace realloc by default
Adds realloc to the list of heap functions traced by -record_heap by
drmemtrace by default.  Also fills in missing variants for tc_ and
__libc for calloc.

Adds a sanity check to the burst_malloc test that we find realloc, and
extends burst_malloc to post-process and walk the trace and confirm we
actually see markers for realloc.  Reduces the iteration count to try
to keep the test from taking too long in the suite.

Fixes #4465