i#3544 RV64: Fix invalid instr width handling when decoding
Currently, valid RISC-V instruction length can be 16-bit (compressed instruction) or 32-bit, but according to the "Extended Instruction-Length Encoding" section of the RISC-V Unprivileged Specification 20191213 on page 8, instruction length can be greater than 32-bit in the future.

Our RISC-V instruction_width() function follows the spec except that it returns the wrong length for the last case. While this patch fixes that, it may still return an unsupported length for bad data.

While performing decoding, sometimes we decode bad data and expect the decoder to print an error message and carry on. Therefore, we should not assert unsupported instruction length, this patch also removes the corresponding assert in decode_common().

Issue: #3544