i#5036 A64 scatter/gather: Fix drcachesim offline tracing
A couple of places in the drcachesim tracer there was code that handles
scatter/gather instructions that was guarded by #ifdef X86 which meant
offline traces that include scatter/gather instructions were producing
errors on AArch64.

For example running the aarch64 scattergather test app:
    $bin64/drrun -t drcachesim -offline -- suite/tests/bin/client.drx-scattergather
    $bin64/drrun -t drcachesim -indir drmemtrace.client.drx-scattergather.662655.1769.dir

produced the error:
    ERROR: failed to initialize analyzer: raw2trace failed: Failed to
    process file for thread 662655: memref entry found outside of bb

This commit enables the code on AArch64 as well and fixes the error.

Issue: #5036