i#6364: add start_instruction to schedule entry comparator. (#6366)
We have captured traces from the same thread with the same CPU and
timestamp, their start_instructions are tens of instructions apart.

The current comparators in raw2trace and the invariant checker do not
take start_instruction into account. And it leads to indeterministic
sorting results. It causes the invariant checker to report cpu/serial
entry does not match trace error when raw2trace and the invariant
checker sort the entries differently.

The change is to compare the start_instruction when two schedule entries
have the same CPU, thread ID, and timestamp.
And add a unit test to cover the case where two schedule entries have
the same CPU, thread ID, and timestamp, but different start_instruction.

Fixes: #6364