i#6394: Replace sentinel in trace with real return address. (#6395)
Update drwrap to check if the return address is a sentinel. If it is,
look for the real return address at the outer level.

When -record_replace_retaddr is used in X86 platforms, the return
address in the stack is replaced by the sentinel.
For a tailcall, the current implementation uses the address in the
stack, the sentinel, as the return address. The change is to check if
the return address in the stack is the sentinel or not. If it is,
replace it with the return address of the outer level.

The problem can be reproduced by 

./build/bin64/drrun -t drcachesim -offline -record_replace_retaddr
-record_function 'tailcall_with_retaddr|1&foo|1' --
../build_suite/build_debug-internal-64/suite/tests/bin/common.getretaddr

A test is added to cover -record_replace_retaddr with common.getretaddr
which has a tailcall.

Fixes: #6394