i#5843 scheduler: Add direct thread switch support
Adds support for the TRACE_MARKER_TYPE_DIRECT_THREAD_SWITCH marker,
when it appears after TRACE_MARKER_TYPE_MAYBE_BLOCKING_SYSCALL.  The
scheduler directly switches to the target thread if it is on the ready
queue.  Performing a forced migration if the target is running on
another output is not yet implemented.  Once i/o wait states are
added, waking up a target thread will be added, but that is future
work as well.

Adds a simple unit test.

Issue: #5843