i#5843 scheduler: Add flexible_queue_t and use it in scheduler_t
Adds a flexible priority queue class which tracks indices and so
supports asking whether an entry is in the queue and removing an entry
from anywhere in the queue.  Adds a simple unit test.

Changes the scheduler to use this new queue class, in anticipation of
needing both new features to handle direct targeted thread switches.

Issue: #5843