i#6397: Add instr_is_opnd_store_source() (#6405)
Adds a new IR API routine instr_is_opnd_store_source() which identifies
whether a given source operand supplies the value to be stored into
memory for a store instruction. Although generally DR avoids modeling
dataflow, especially for complex SIMD shifts, this data identification
is useful enough that we provide a helper function. However, we do not
go as far as annotating the decoder operands, and instead deduce the
answer based on conventions in our IR, which should suffice for now and
even if not as clean as decoder markup it is better to have this helper
in one place we could improve in the future instead of different users
trying to figure this out.

Adds some tests for each architecture.

Fixes #6397