i#3544 RV64: Implement insert_reachable_cti
Implements insert_reachable_cti(), which is a core procedure used by the dispatcher, meta call, etc. Similar to AArch64, we always use an indirect branch to do the jump, so this procedure always returns true.

Also removed is a related FIXME: the `CALL_SCRATCH_REG` is used as a scratch to hold the indirect branch target, using t6 (t for tmp, i.e. caller saved) should be safe.