i#5365: Build core unit tests with SVE enabled
Build most core tests with SVE flags and high optimisation (-O3), if building
on an AARCH64 SVE machine.

Tests which fail when built with -O3 are not included.

Add some error checking to a few tests to allow the -O3 build and update
template (expected output) files as necessary.

Issue #6429 raised to cover making the removal of optimization flags more
granular.

Issue: #5365