i#3544 RV64: Fill in the missing pieces to run actual programs
This patch fills in many missing pieces related to RISC-V porting.

Although our philosophy is to keep patches as small as possible, IMHO these pieces would be painful to submit individually.

With this patch, plus one more patch to mangle.c yet to be submitted, we are finally able to run some programs, which means some tests can be enabled/ported for RISC-V and more RISC-V specific tests can be written.